+++ src/agnocastlib/src/agnocast_utils.cpp
@@ -2,24 +2,16 @@
 
 #include "agnocast/node/agnocast_node.hpp"
 
-#include <rclcpp/version.h>
 #include <sys/stat.h>
 
-#include <algorithm>
-#include <cctype>
 #include <cerrno>
 #include <cstdlib>
 #include <cstring>
 #include <limits>
-#include <memory>
-#include <sstream>
-#include <string_view>
 #include <system_error>
-#include <tuple>
 
 namespace agnocast
 {
-
 rclcpp::Logger logger = rclcpp::get_logger("Agnocast");
 bool is_bridge_process = false;
 
@@ -142,159 +134,6 @@ const void * get_node_base_address(rclcpp::Node * node)
 {
   return static_cast<const void *>(
     node->get_node_base_interface()->get_shared_rcl_node_handle().get());
-}
-
-// === Begin code from rclcpp ===
-// Copyright 2024 Sony Group Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// This file has been modified from the original.
-
-#if RCLCPP_VERSION_MAJOR < 28
-
-namespace
-{
-
-std::string string_trim(std::string_view str_v)
-{
-  const auto * begin =
-    std::find_if_not(str_v.begin(), str_v.end(), [](unsigned char ch) { return std::isspace(ch); });
-  const auto * end = std::find_if_not(str_v.rbegin(), str_v.rend(), [](unsigned char ch) {
-                       return std::isspace(ch);
-                     }).base();
-  if (begin >= end) {
-    return {};
-  }
-  return {begin, end};
-}
-
-std::tuple<std::string, std::string, std::string> extract_type_identifier(
-  const std::string & full_type)
-{
-  char type_separator = '/';
-  auto sep_position_back = full_type.find_last_of(type_separator);
-  auto sep_position_front = full_type.find_first_of(type_separator);
-  if (
-    sep_position_back == std::string::npos || sep_position_front == 0 || sep_position_back == 0 ||
-    sep_position_back == full_type.length() - 1) {
-    throw std::runtime_error(
-      "Message type is not of the form package/type and cannot be processed");
-  }
-
-  std::string package_name = full_type.substr(0, sep_position_front);
-  std::string middle_module;
-  if (sep_position_back - sep_position_front > 0) {
-    middle_module =
-      full_type.substr(sep_position_front + 1, sep_position_back - sep_position_front - 1);
-  }
-  std::string type_name = full_type.substr(sep_position_back + 1);
-
-  return std::make_tuple(
-    string_trim(package_name), string_trim(middle_module), string_trim(type_name));
-}
-
-const void * get_typesupport_handle_impl(
-  const std::string & type, const std::string & typesupport_identifier,
-  const std::string & typesupport_name, const std::string & symbol_part_name,
-  const std::string & middle_module_additional, rcpputils::SharedLibrary & library)
-{
-  std::string package_name;
-  std::string middle_module;
-  std::string type_name;
-  std::tie(package_name, middle_module, type_name) = extract_type_identifier(type);
-
-  if (middle_module.empty()) {
-    middle_module = middle_module_additional;
-  }
-
-  auto mk_error = [&package_name, &type_name, &typesupport_name](auto reason) {
-    std::stringstream rcutils_dynamic_loading_error;
-    rcutils_dynamic_loading_error << "Something went wrong loading the typesupport library for "
-                                  << typesupport_name << " type " << package_name << "/"
-                                  << type_name << ". " << reason;
-    return rcutils_dynamic_loading_error.str();
-  };
-
-  try {
-    std::string symbol_name = typesupport_identifier + symbol_part_name + package_name + "__" +
-                              middle_module + "__" + type_name;
-    const void * (*get_ts)() = nullptr;
-    // This will throw runtime_error if the symbol was not found.
-    get_ts = reinterpret_cast<decltype(get_ts)>(library.get_symbol(symbol_name));
-    return get_ts();
-  } catch (std::runtime_error &) {
-    throw std::runtime_error{mk_error("Library could not be found.")};
-  }
-}
-
-const rosidl_service_type_support_t * get_service_typesupport_handle(
-  const std::string & type, const std::string & typesupport_identifier,
-  rcpputils::SharedLibrary & library)
-{
-  static const std::string typesupport_name = "service";
-  static const std::string symbol_part_name = "__get_service_type_support_handle__";
-  static const std::string middle_module_additional = "srv";
-
-  return static_cast<const rosidl_service_type_support_t *>(get_typesupport_handle_impl(
-    type, typesupport_identifier, typesupport_name, symbol_part_name, middle_module_additional,
-    library));
-}
-
-}  // namespace
-
-#endif
-// === End code from rclcpp ===
-
-ServiceTsBundle load_service_typesupport(const std::string & service_type)
-{
-  static const std::string ts_identifier = "rosidl_typesupport_cpp";
-  static const std::string ts_identifier_introspection = "rosidl_typesupport_introspection_cpp";
-
-  const std::string request_type = service_type + "_Request";
-  const std::string response_type = service_type + "_Response";
-
-  ServiceTsBundle bundle;
-
-  bundle.ts_lib = rclcpp::get_typesupport_library(service_type, ts_identifier);
-  bundle.ts_lib_introspection =
-    rclcpp::get_typesupport_library(service_type, ts_identifier_introspection);
-
-  // Not the service introspection handle: its ServiceMembers is filled in lazily on first access
-  // without synchronization.
-#if RCLCPP_VERSION_MAJOR >= 28
-  bundle.service_ts =
-    rclcpp::get_service_typesupport_handle(service_type, ts_identifier, *bundle.ts_lib);
-
-  const rosidl_message_type_support_t * request_ts = rclcpp::get_message_typesupport_handle(
-    request_type, ts_identifier_introspection, *bundle.ts_lib_introspection);
-  const rosidl_message_type_support_t * response_ts = rclcpp::get_message_typesupport_handle(
-    response_type, ts_identifier_introspection, *bundle.ts_lib_introspection);
-#else
-  bundle.service_ts = get_service_typesupport_handle(service_type, ts_identifier, *bundle.ts_lib);
-
-  const rosidl_message_type_support_t * request_ts = rclcpp::get_typesupport_handle(
-    request_type, ts_identifier_introspection, *bundle.ts_lib_introspection);
-  const rosidl_message_type_support_t * response_ts = rclcpp::get_typesupport_handle(
-    response_type, ts_identifier_introspection, *bundle.ts_lib_introspection);
-#endif
-
-  bundle.request_members =
-    static_cast<const rosidl_typesupport_introspection_cpp::MessageMembers *>(request_ts->data);
-  bundle.response_members =
-    static_cast<const rosidl_typesupport_introspection_cpp::MessageMembers *>(response_ts->data);
-
-  return bundle;
 }
 
 }  // namespace agnocast
